requirements.txt
Flask
Flask-Cors
torch
timm
albumentations
opencv-python-headless
Pillow
gunicorn